Moved generation logic into Area Model
