--- util/sample_generator.py
@@ -1,206 +1,0 @@
-import random
-
-
-class Rect:
-    ''' 
-    Rects are created in order to manage regions that
-    represent rooms. They are also for checking to see if rooms
-    intersect when being generated and placed on the map
-    '''
-
-    def __init__(self, coords, size):
-        '''
-        coords is a tuple containing the top left x, y coordinates
-        of the rect. size is a tuple containing the height and width
-        of the rect.
-        '''
-        self.x1, self.y1 = coords
-        self.w, self.h = size
-        self.x2 = self.x1 + self.w
-        self.y2 = self.y1 + self.h
-
-    @property
-    def center(self):
-        ''' 
-        The coordinates of the center of the rectangle in order
-        to create tunneling targets for connecting rects together.
-        '''
-        center_x = int((self.x1 + self.x2) / 2)
-        center_y = int((self.y1 + self.y2) / 2)
-
-        return (center_x, center_y)
-
-    def intersects(self, target):
-        '''
-        Returns true if the rects intersect on both the x and y plane.
-        '''
-        intersect = self.x1 <= target.x2 and self.x2 >= target.x1 and self.y1 <= target.y2 and self.y2 <= target.y1
-
-        return intersect
-
-
-class Room:
-    def __init__(self, sector_type, x, y):
-        self.sector_type = sector_type
-        self.n_to = None
-        self.s_to = None
-        self.e_to = None
-        self.w_to = None
-        self.x = x
-        self.y = y
-
-    def __repr__(self):
-        if self.e_to is not None:
-            return f"({self.x}, {self.y}) -> ({self.e_to.x}, {self.e_to.y})"
-        return f"({self.x}, {self.y})"
-
-    def connect_rooms(self, connecting_room, direction):
-        '''
-        Connect two rooms in the given n/s/e/w direction
-        '''
-        reverse_dirs = {"n": "s", "s": "n", "e": "w", "w": "e"}
-        reverse_dir = reverse_dirs[direction]
-        setattr(self, f"{direction}_to", connecting_room)
-        setattr(connecting_room, f"{reverse_dir}_to", self)
-
-    def get_room_in_direction(self, direction):
-        '''
-        Connect two rooms in the given n/s/e/w direction
-        '''
-        return getattr(self, f"{direction}_to")
-
-
-class Map:
-    def __init__(self):
-        self.grid = None
-        self.width = 0
-        self.height = 0
-        self.max_regions = 30
-        self.max_region_size = 7
-        self.min_region_size = 3
-
-    def draw_region(self, region):
-        '''
-        Region is a rect instance. We iterate over all coordinates within the
-        bounds of the rect and populate those coordinates with new instances
-        of Room that have floor sector type.
-        '''
-        for x in range(region.x1, region.x2):
-            for y in range(region.y1, region.y2):
-                self.grid[x][y] = Room('0', x, y)
-
-    def draw_tunnel(self, a, b):
-        '''
-        a and b are the centermost points of two seperate regions that have
-        been drawn on the map grid. We dig out a tunnel, by making a room
-        with the sector type of floor between the x and then the y coords
-        of each region.
-        '''
-        x1, y1 = a
-        x2, y2 = b
-
-        for x in range(min(x1, x2), max(x1, x2) + 1):
-            self.grid[x][y1] = Room('0', x, y1)
-
-        for y in range(min(y1, y2), max(y1, y2) + 1):
-            self.grid[x2][y] = Room('0', x1, y)
-
-    def generate_regions(self, size_x, size_y):
-        '''
-        Generates a maximum number of regioins equal to the defined value
-        in the World class. Region are generated by creating a rect and
-        attempting to place it on the map grid. If the rect does not
-        intersect with any existing region, it then draws a new region
-        on the map and creates a tunnel between it and the last drawn
-        region.
-        '''
-
-        self.grid = []
-        self.width = size_x
-        self.height = size_y
-
-        '''
-        Create a 2d array and populate each element with a Room object that
-        has a wall sector_type.
-        '''
-        for x in range(0, self.width):
-            self.grid.append([])
-            for y in range(0, self.height):
-                self.grid[x].append(Room('1', x, y))
-
-        '''
-        Store the rect of each created region in a list so that we can see
-        if future rects intersect and should not be placed.
-        '''
-        region_list = []
-
-        '''
-        Remember the center of the last region placed so we have a target
-        to draw a tunnel toward when placing the next region.
-        '''
-        center_of_last_region = None
-
-        '''
-        We loop a number of times equal to the maximum allowed regions on the
-        map. This does not ensure that many regions will be created.
-        '''
-
-        for i in range(self.max_regions):
-
-            '''
-            Randomize a width, height and x,y coordinates for a new region
-            '''
-            w = random.randint(self.min_region_size, self.max_region_size)
-            h = random.randint(self.min_region_size, self.max_region_size)
-            x = random.randint(2, self.width-(w+2))
-            y = random.randint(2, self.height-(h+2))
-
-            region = Rect((x, y), (w, h))
-
-            '''
-            Track if we were able to place the newly created region.
-            '''
-            placement_failed = False
-
-            '''
-            Iterate over previously created regions and see if the
-            new region intersects those. If it does, placement has
-            failed and we start over generating a new region.
-            '''
-            for placed_region in region_list:
-                if region.intersects(placed_region):
-                    placement_failed = True
-                    break
-
-            '''
-            If placement succeeds we append our region to the list of
-            placed regions, draw it onto our map. If we have stored
-            the center of a previously placed region, we will draw a 
-            tunnel between the new and last one. We then set last center
-            to that of the newly placed region.
-            '''
-            if not placement_failed:
-                region_list.append(region)
-                self.draw_region(region)
-                if center_of_last_region is not None:
-                    self.draw_tunnel(region.center, center_of_last_region)
-
-                center_of_last_region = region.center
-
-    def print_rooms(self):
-        '''
-        Print out rooms in a series of 1, 0 based on sector.
-        '''
-        string = ""
-        for i in range(0, self.width):
-            string += "\n"
-            for j in range(0, self.height):
-                string += f"{self.grid[i][j].sector_type}"
-        print(string)
-
-
-m = Map()
-width = 50
-height = 50
-m.generate_regions(width, height)
-m.print_rooms()
